requirements.txt
# Core
python-dotenv>=1.0.0
pyyaml>=6.0.1
pydantic>=2.6

# Data
pandas>=2.1
numpy>=1.26

# LLM (Gemini – supported)
google-genai>=0.4.0

# Plotting / reporting
matplotlib>=3.8
seaborn>=0.13

# CLI / utils
tqdm>=4.66

# Testing
pytest>=8.0
pytest-mock>=3.12

tabulate>=0.9
markdown==3.10
openpyxl==3.1.

faiss-cpu==1.13.25